This is the finished game
The code prompts the user of what button color to click next in the sequence. It is up to the user to keep up and remember the previous colors to click. The rounds level up each time the user is successful and the code keeps track of the user's button clicks, compares them to what the game has prompted, and determines whether or not the user has been successful or gotten them wrong. If the user is wrong, the game is over and can be restarted.